pyproject.toml
[project]
name = "adk-customer-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "google-adk>=1.18.0",
    "litellm>=1.80.0",
]
